requirements-test.txt
-r requirements.txt

pytest-mock ~= 3.14

coverage ~= 7.3.0